backend/requirements.txt
fastapi>=0.100
uvicorn[standard]>=0.23
aiofiles>=23.0
jinja2>=3.1
beautifulsoup4>=4.12
requests>=2.31
pyusb>=1.2
psutil>=5.9
python-multipart>=0.0.6
websockets>=12.0
wsproto>=1.2
pydantic>=2.0
